pyproject.toml
[tool.poetry]
name = "example_template"
version = "1.0.0"
description = "This is a template example for python packages"
authors = ["samet <samettaspinarx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
langchain = "^0.0.325"
python-dotenv = "^1.0.0"
pandas = "^2.1.2"
openai = "^0.28.1"
tiktoken = "^0.5.1"
tqdm = "^4.66.1"
ruff = "0.1.3"
isort = "^5.12.0"
black = "^23.10.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ["py38"]
exclude = '''
/(
    \.git
  | \.venv
  | \.direnv
  | _build
  | build
  | dist
  | \.env
  | \.vscode
  | \.cache
  | github
  | __pycache__
  | .cloudignore
)/
'''

[tool.isort]
py_version = "38"
line_length = 80
lines_after_imports = 2
lines_between_sections = 1
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = "true"
combine_as_imports = "true"
use_parentheses = "true"
skip = ["./docs", ".direnv"]
skip_glob = ["*/__init__.py"]
skip_gitignore = "true"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"

[tool.pytest.ini_options]
minversion = "6.1"
addopts = "--cov"
testpaths = [
    "tests_unit",
]
